+++ src/main.java
@@ -1,0 +1,9 @@
+import controller.GameController;
+
+public class main {
+
+    public static void main(String[] args) {
+        new GameController();
+    }
+
+}
