+++ src/model/factories/EntityFactory.java
@@ -1,0 +1,4 @@
+package model.factories;
+
+public class EntityFactory {
+}
