+++ src/model/factories/TerrainFactory.java
@@ -1,0 +1,4 @@
+package model.factories;
+
+public class TerrainFactory {
+}
